+++ api/client/image/search.go
@@ -52,8 +52,8 @@ func NewSearchCommand(dockerCli *client.DockerCli) *cobra.Command {
 	flags.BoolVar(&opts.automated, "automated", false, "Only show automated builds")
 	flags.UintVarP(&opts.stars, "stars", "s", 0, "Only displays with at least x stars")
 
-	flags.MarkDeprecated("automated", "Use --filter=automated=true instead")
-	flags.MarkDeprecated("stars", "Use --filter=stars=3 instead")
+	flags.MarkDeprecated("automated", "use --filter=automated=true instead")
+	flags.MarkDeprecated("stars", "use --filter=stars=3 instead")
 
 	return cmd
 }
